starting to create app file structure, added namespaces, somehow broke make...
