+++ src/main.cpp
@@ -1,30 +1,12 @@
 #include <gtkmm.h>
-#include "../include/pfs/procfs.hpp"
-#include <iostream>
+#include "../include/gsm/app.hpp"
 
-using namespace pfs;
-using namespace std;
+namespace gsm {
 
-Glib::RefPtr<Gtk::Application> app;
+    int main(int argc, char* argv[]) {
 
-void load_ui() {
-    auto builder = Gtk::Builder::create_from_file("./glade/glade.ui");
-    Gtk::Window *window;
-    builder->get_widget("main_window", window);
-    app->add_window(*window);
-    window->show();
-}
-
-int main(int argc, char* argv[]) {
-
-    Glib::init();
-    app = Gtk::Application::create("com.github.mdmellor.gsystem-manager");
-    app->signal_activate().connect(sigc::ptr_fun(&load_ui));
-    procfs p = procfs();
-    for (auto &x : p.get_processes()) {
-        cout << x.get_comm() << endl;
+        Glib::init();
+        return start_app(argc, argv);
     }
-    return app->run(argc, argv);
 }
-
 // Compile by executing "make" from root project directory
